- replaced with u32 for 64bit architecture safe.
    94c978c437b9c01a4f562eecbf12d91f4cb5126e
1.0.0pre2
    3ddbc85a141090a34a25b224b0f0e52a7d149636

+++ alsa/pci/au88x0/au88x0.h
@@ -141,7 +141,7 @@ struct snd_vortex {
 
 	/* PCI hardware resources */
 	unsigned long io;
-	unsigned long *mmio;
+	u32 *mmio;
 	unsigned int irq;
     spinlock_t lock;
